% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmPQL.mod.R
\name{glmmPQL.mod}
\alias{glmmPQL.mod}
\title{Fit Generalized Linear Mixed Models via PQL}
\usage{
glmmPQL.mod(fixed, random, family, data, correlation, weights, control,
  niter = 50, verbose = TRUE)
}
\arguments{
\item{...}{See \code{\link[MASS:glmmPQL]{glmmPQL}} for details.}
}
\description{
Fit a GLMM model with multivariate normal random effects, using Penalized Quasi-Likelihood.
This function is a minor modification of the \code{\link[MASS:glmmPQL]{glmmPQL}} function to
a) use REML (instead of ML) for estimation,
b) provide necessary output to standardize responses for testing, and
c) increase the default max iterations to 50 (from 10).
Otherwise, the functions are identical.
}
\seealso{
\code{\link[MASS:glmmPQL]{glmmPQL}}, \code{\link[nlme:lme]{lme}}
}
