% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std.glmmPQL.R
\name{std.glmmPQL}
\alias{std.glmmPQL}
\title{Update function call with standardized data}
\usage{
std.glmmPQL(fit.glmmPQL)
}
\arguments{
\item{fit.glmmPQL}{Model estimate from glmmPQL.mod.}
}
\value{
Returns an updated function call for PQL estimation.
}
\description{
Updates the function call from glmmPQL.mod so the working LMM has
standardized residuals, for testing. The normalized response and design matrices
are all standardized accordingly.
}
