% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.asaRLRT.R
\name{test.asaRLRT}
\alias{test.asaRLRT}
\title{Asymptotic-approximate RLRT for variance components in GLMMs}
\usage{
test.asaRLRT(fit.glmmPQL, nsim = 10000)
}
\arguments{
\item{fit.glmmPQL}{Alternative model estimated using the \code{glmmPQL.mod} function.}
}
\value{
Returns a list of testing results:
\itemize{
\item \code{asaRLRT}: outcome of the asaRLRT test (statistic and p-value)
\item \code{std.data}: data.frame of standardized normalized responses (Ytilde),
fixed effects design, and random effects design
\item \code{fit.alt}: lme under the alternative hypothesis
\item \code{fit.null}: lme under the null hypothesis
}
}
\description{
Conducts an asymptotic-approximate restricted-likelihood ratio test (as-aRLRT)
for random effects (variance components) in generalized linear mixed models (GLMMs).
The method applies RLRT to the working linear mixed model (LMM) approximation
for the GLMM used in PQL estimation. The test statistic is compared to the
asymptotic null distribution from Self and Liang (1987).
}
\details{
Compared to the recommended aRLRT, this method instead uses an asymptotic null
distribution (Self and Liang 1987), which typically has conservative type I error
rates and lower power. The function assumes that the first specified random effect is
the factor of interest.
This random effect must be independent of all other random effects, but can have any
covariance structure.
Please use the \code{glmmPQL.mod} function in this package; it has minor modifications
from \code{MASS::glmmPQL}.
NOTE: The glmmPQL model cannot use the ~(1|group) notation
for any random effects (fixed are ok). Please specify a random interept directly
by adding an intercept column (column of 1s) and using ~(0+'name'|group). Variable name
does not matter.

The function does not currently support simultaneous testing of fixed and random effects.
That may be added in the future.
}
\examples{
\dontrun{
library(glmm)
data(salamander) # load data
salamander$int <- 1 # add column of 1s for random intercept
# Fit alternative hypothesis
glmm.fit <- glmmPQL.mod(Mate~0+Cross, random=list(~0+int|Male),
data=salamander, family=binomial, weights=rep(1,nrow(salamander)))
# Test significance of random subject-specific intercept for Male salamanders
asaRLRT <- test.asaRLRT(glmm.fit)
}

}
\seealso{
\code{\link{test.aRLRT}}
}
